Core: Refactor
