+++ src/types/aggregation/aggregation.ts
@@ -1,7 +1,7 @@
 import { AggregationCursor, Document } from 'mongodb'
-import type { Collstats } from '../colstats'
+import type { Collstats } from '.'
+import { GeoNear } from '.'
 import { TsFilter } from '../filter'
-import { GeoNear } from '../geoNear'
 import { TsProjection } from '../projection'
 import { TsSort } from '../sort'
 import { RemodelType } from '../util'
