+++ src/types/aggregation/index.ts
@@ -1,1 +1,3 @@
 export * from './aggregation'
+export * from './colstats'
+export * from './geoNear'
